pyproject.toml
[project]
name = "ronin"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "starlette>=0.52.1",
    "uvicorn[standard]>=0.40.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.0",
    "httpx>=0.28.1",
    "pyyaml>=6.0.0",
    "mahjong==2.0.0.dev1",
    "msgpack>=1.1.2",
    "xiangting>=5.0.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
    "bcrypt>=4.0.0",
    "anyio>=4.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-timeout>=2.4.0",
    "ruff>=0.15.1",
    "ty>=0.0.17",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/lobby", "backend/game", "backend/shared"]

[tool.pytest.ini_options]
testpaths = ["backend"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
python_files = "test_*.py"
addopts = "-p no:warnings --no-header --tb=short -qq --cov --cov-report=term:skip-covered"
timeout = 10

[tool.ruff]
src = ["backend"]
target-version = "py314"
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D", # Dockstring styles
    "EM101",  # Exception must not use a string literal
    "EM102",  # Exception must not use an f-string literal
    "TRY003", # Avoid specifying long messages outside exception class
    "G004",  # Convert to lazy `%` formatting
    "TD", # TODO comments
    "FIX002",# Line contains TODO
]

[tool.ruff.lint.per-file-ignores]
"{test_*.py,conftest.py,**/tests/**/helpers.py,**/tests/**/mocks/*.py}" = [
    "ANN", # Missing type annotation
    "ARG", # Unused function arguments
    "PLR0913", # Too many arguments (test factories)
    "PLR0915", # Too many statements
    "PLR2004", # Magic value used in comparison
    "S101", # Use of `assert` detected
    "S105", # Possible hardcoded password
    "S106", # Possible hardcoded password assigned to argument
    "SLF001", # Private member accessed
]
# Base class methods define the interface contract; parameters are unused in the default implementation
"backend/game/logic/ai_player.py" = ["ARG002"]
"backend/game/logic/shanten.py" = ["TID251"]

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = ["pydantic.BaseModel", "pydantic_settings.BaseSettings"]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"mahjong.shanten".msg = "Use game.logic.shanten instead of mahjong.shanten directly."

[tool.ty.environment]
python-version = "3.14"

[tool.ty.src]
include = ["backend"]

[tool.ty.rules]
# Promote all non-error default rules to error for stricter checking
division-by-zero = "error"               # default: ignore
possibly-unresolved-reference = "error"   # default: warn
possibly-missing-attribute = "error"      # default: warn
possibly-missing-import = "error"         # default: warn
possibly-missing-implicit-call = "error"  # default: warn
deprecated = "error"                      # default: warn
redundant-cast = "error"                  # default: warn
ineffective-final = "error"              # default: warn
ambiguous-protocol-member = "error"       # default: warn
invalid-ignore-comment = "error"          # default: warn
ignore-comment-unknown-rule = "error"     # default: warn
invalid-legacy-positional-parameter = "error"  # default: warn
unused-ignore-comment = "error"           # default: warn
unused-type-ignore-comment = "error"      # default: warn

[[tool.ty.overrides]]
include = ["**/test_*.py", "**/conftest.py"]

[tool.ty.overrides.rules]
# relax rules for test files - tests often use dynamic fixtures, mocks,
# and untyped external libraries (mahjong) that produce Unknown types
# Note: invalid-argument-type is ignored because Pydantic validators coerce
# list -> tuple at runtime, making tests functionally correct
# Note: possibly-missing-attribute and unresolved-attribute are ignored because
# tests often access union type members after type-narrowing assertions that ty doesn't track
# Note: missing-argument is ignored because pydantic-settings classes read
# required fields from environment variables at runtime
possibly-unresolved-reference = "warn"
possibly-missing-attribute = "ignore"
unresolved-attribute = "ignore"
invalid-argument-type = "ignore"
missing-argument = "ignore"
invalid-parameter-default = "warn"
unused-ignore-comment = "ignore"

[tool.coverage.run]
source = ["backend"]
omit = ["**/tests/*", "**/test_*", "backend/game/logic/ai_player.py", "backend/game/logic/ai_player_controller.py", "backend/debug.py", "backend/shared/logging.py"]

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
    '^\\s*\\.\\.\\.\\s*$',
    '^\\s*pass\\s*$',
    "@(abc\\.)?abstractmethod",
]
fail_under = 100
